+++ src/main/java/ao/adoop/mapreduce/Configuration.java
@@ -10,11 +10,6 @@ public class Configuration {
 	
 	public Path mapOutputBufferDir = Paths.get(systemBaseDir.toString(), "map-output-buffer");
 	public String mapOutputFileExtension = ".csv";
-	
-//	public Path reduceOutputBufferDir = Paths.get(systemBaseDir.toString(), "reduce-output-buffer");
-//	public Path namedReduceOutputBaseDir = Paths.get(systemBaseDir.toString(), "named-reduce-output-buffer");
-//	public String reduceOutputFileName = "output";
-//	public String reduceOutputFileExtension = ".csv";
 		
 	public Path finalOutputDir = Paths.get(systemBaseDir.toString(), "final-outputs");
 	public String finalOutputFileName = "output";
@@ -26,6 +21,8 @@ public class Configuration {
 
 	public String generateReduceOutputFileName(String workerId) {
 		return "part-r-" + workerId + this.finalOutputFileExtension;
-	}
+	};
 
+	public long threadMaxThreashhold = 30;
+	public String threadMaxThreashholdUnit = "MB";
 }
