--- src/test/java/ao/adoop/test/test_usermodules/InvalidClass.java
@@ -1,5 +1,0 @@
-package ao.adoop.test.test_usermodules;
-
-public class InvalidClass {
-
-}
