+++ src/test/java/ao/adoop/test/unit/DataLoaderTest.java
@@ -1,4 +1,4 @@
-package adooptest;
+package ao.adoop.test.unit;
 
 import static org.junit.jupiter.api.Assertions.*;
