+++ src/test/java/ao/adoop/test/utils/usermodules/InvalidClass.java
@@ -1,0 +1,5 @@
+package ao.adoop.test.utils.usermodules;
+
+public class InvalidClass {
+
+}
