+++ src/test/java/ao/adoop/test/utils/usermodules/MapperForIntegrationTest2.java
@@ -1,4 +1,4 @@
-package ao.adoop.test.test_usermodules;
+package ao.adoop.test.utils.usermodules;
 
 import java.io.File;
 
@@ -6,9 +6,10 @@ import ao.adoop.mapreduce.Configuration;
 import ao.adoop.mapreduce.Context;
 import ao.adoop.mapreduce.Mapper;
 
-public class MapperForIntegrationTest1 extends Mapper {
+public class MapperForIntegrationTest2 extends Mapper {
 
-	public MapperForIntegrationTest1(String workerId, Configuration config, File inputFile,
+
+	public MapperForIntegrationTest2(String workerId, Configuration config, File inputFile,
 			int startIndex, int endIndex, String[] addedNamedOutputs) {
 		super(workerId, config, inputFile, startIndex, endIndex, addedNamedOutputs);
 	}
@@ -16,7 +17,7 @@ public class MapperForIntegrationTest1 extends Mapper {
 	@Override
 	public void map(String key, String value, Context context) {
 		String[] parts = value.split(",");
-		context.write(parts[1], parts[2]);
+		context.write(parts[2], parts[3]);
 	}
 
 }
