+++ src/test/java/ao/adoop/test/utils/usermodules/MultipleOutputReducerForIntegrationTest.java
@@ -1,4 +1,4 @@
-package ao.adoop.test.test_usermodules;
+package ao.adoop.test.utils.usermodules;
 
 import java.io.File;
 import java.util.ArrayList;
