+++ src/test/java/ao/adoop/test/utils/usermodules/UnitTestMapper.java
@@ -1,4 +1,4 @@
-package ao.adoop.test.test_usermodules;
+package ao.adoop.test.utils.usermodules;
 
 import java.io.File;
 
@@ -6,16 +6,16 @@ import ao.adoop.mapreduce.Configuration;
 import ao.adoop.mapreduce.Context;
 import ao.adoop.mapreduce.Mapper;
 
-public class UnitTestMapper2 extends Mapper{
+public class UnitTestMapper extends Mapper{
 
-	public UnitTestMapper2(String workerId, Configuration config, File inputFile, Integer startIndex,
+
+	public UnitTestMapper(String workerId, Configuration config, File inputFile, Integer startIndex,
 			Integer endIndex, String[] addedNamedOutputs) {
 		super(workerId, config, inputFile, startIndex, endIndex, addedNamedOutputs);
 	}
 
-	@Override
 	public void map(String key, String value, Context context) {
-		context.write("Key", "Value");
-	}
-
+		String[] arrOfStr = value.split("/", 3);
+		context.write(arrOfStr[1], arrOfStr[2]);
+	};
 }
